Level 2

+++ Assets/Scripts/Interfaces/IRage.cs
@@ -2,6 +2,6 @@
 {
     public interface IRage
     {
-        public int GetRageValue();
+        public float GetRageValue();
     }
 }
